solve all

+++ exercises/enums/enums2.rs
@@ -5,10 +5,11 @@
 
 #[derive(Debug)]
 enum Message {
-    Move {x:i32,y:i32},
-    Echo (String),
-    ChangeColor (i32,i32,i32),
-    Quit
+    // TODO: define the different variants used below
+    Move { x: u8, y: u8 },
+    Echo(String),
+    ChangeColor(u8, u8, u8),
+    Quit,
 }
 
 impl Message {
